feat(service): delete function check if id is valid and if the result is not null

+++ src/app/service/ProductService.ts
@@ -8,9 +8,7 @@ import isValidUuid from '../utils/isValidUuid'
 class ProductService {
   async create (payload: IProductCreate): Promise<IProductResponse> {
     const result = await ProductRepository.create(payload)
-
     if (result === null) throw new BadRequestError('Product not created')
-
     return result
   }
 
@@ -23,9 +21,7 @@ class ProductService {
     queryBuilded.stock_control_enabled = true
 
     const result: PaginateResult<IProductResponse> = await ProductRepository.findAll(queryBuilded, page ?? 1)
-
     if (result.totalCount === 0) throw new NotFoundError('Product not found')
-
     return result
   }
 
@@ -33,9 +29,7 @@ class ProductService {
     if (!isValidUuid(id)) throw new BadRequestError('Product id is not valid')
 
     const result = await ProductRepository.findOne(id)
-
     if (result === null) throw new NotFoundError('Product not found')
-
     return result
   }
 
@@ -43,7 +37,6 @@ class ProductService {
     const result: PaginateResult<IProductResponse> = await ProductRepository.findLowStock(page ?? 1)
 
     if (result.totalCount === 0) throw new NotFoundError('Product not found')
-
     return result
   }
 
@@ -53,14 +46,15 @@ class ProductService {
     payload.qtd_stock === 0 ? payload.stock_control_enabled = false : payload.stock_control_enabled = true
 
     const result = await ProductRepository.update(id, payload)
-
     if (result === null) throw new NotFoundError('Product not found')
-
     return result
   }
 
   async delete (id: string): Promise<IProductResponse> {
+    if (!isValidUuid(id)) throw new BadRequestError('Product id is not valid')
+
     const result = await ProductRepository.delete(id)
+    if (result === null) throw new NotFoundError('Product not found')
     return result
   }
 }
